Refactor Task Creation in Backend (#4933)
* [WIP] refactor task creation in backend

* fix taskcontroller

* move remaining methods, rename some things, extract some things

* disable optionget warning again

* remove unused method
