+++ app/controllers/AnnotationIOController.scala
@@ -1,6 +1,6 @@
 package controllers
 
-import java.io.{BufferedOutputStream, File, FileOutputStream, OutputStream}
+import java.io.File
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
@@ -8,9 +8,8 @@ import akka.stream.scaladsl._
 import akka.util.ByteString
 import com.mohiva.play.silhouette.api.Silhouette
 import com.scalableminds.util.accesscontext.{DBAccessContext, GlobalAccessContext}
-import com.scalableminds.util.io.{NamedEnumeratorStream, NamedStream, ZipIO}
+import com.scalableminds.util.io.{NamedEnumeratorStream, ZipIO}
 import com.scalableminds.util.tools.{Fox, FoxImplicits, TextUtils}
-import com.scalableminds.webknossos.datastore.dataformats.wkw.WKWBucketStreamSink
 import com.scalableminds.webknossos.datastore.models.datasource.{AbstractSegmentationLayer, SegmentationLayer}
 import com.scalableminds.webknossos.tracingstore.SkeletonTracing.{SkeletonTracing, SkeletonTracingOpt, SkeletonTracings}
 import com.scalableminds.webknossos.tracingstore.VolumeTracing.{VolumeTracing, VolumeTracingOpt, VolumeTracings}
@@ -29,7 +28,7 @@ import models.team.OrganizationDAO
 import models.user._
 import oxalis.security.WkEnv
 import play.api.i18n.{Messages, MessagesProvider}
-import play.api.libs.Files.{TemporaryFile, TemporaryFileCreator}
+import play.api.libs.Files.TemporaryFile
 import play.api.libs.iteratee.Enumerator
 import play.api.libs.iteratee.streams.IterateeStreams
 import play.api.libs.json.Json
@@ -49,7 +48,6 @@ class AnnotationIOController @Inject()(nmlWriter: NmlWriter,
                                        taskTypeDAO: TaskTypeDAO,
                                        tracingStoreService: TracingStoreService,
                                        annotationService: AnnotationService,
-                                       temporaryFileCreator: TemporaryFileCreator,
                                        sil: Silhouette[WkEnv],
                                        provider: AnnotationInformationProvider,
                                        nmlService: NmlService)(implicit ec: ExecutionContext)
@@ -142,7 +140,7 @@ class AnnotationIOController @Inject()(nmlWriter: NmlWriter,
   private def descriptionForNMLs(descriptions: Seq[Option[String]]) =
     if (descriptions.size == 1) descriptions.headOption.flatten.getOrElse("") else ""
 
-  private def returnError(zipParseResult: NmlResults.ZipParseResult)(implicit messagesProvider: MessagesProvider) =
+  private def returnError(zipParseResult: NmlResults.MultiNmlParseResult)(implicit messagesProvider: MessagesProvider) =
     if (zipParseResult.containsFailure) {
       val errors = zipParseResult.parseResults.flatMap {
         case result: NmlResults.NmlParseFailure =>
