+++ app/models/annotation/AnnotationService.scala
@@ -34,7 +34,7 @@ import models.binary._
 import models.mesh.{MeshDAO, MeshService}
 import models.project.ProjectDAO
 import models.task.{Task, TaskDAO, TaskService, TaskTypeDAO}
-import models.team.{OrganizationDAO, TeamDAO}
+import models.team.{OrganizationDAO, Team, TeamDAO}
 import models.user.{User, UserDAO, UserService}
 import net.liftweb.common.{Box, Full}
 import play.api.i18n.{Messages, MessagesProvider}
@@ -83,8 +83,8 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
     with TextUtils
     with ProtoGeometryImplicits
     with LazyLogging {
-  implicit val actorSystem = ActorSystem()
-  implicit val materializer = ActorMaterializer()
+  implicit val actorSystem: ActorSystem = ActorSystem()
+  implicit val materializer: ActorMaterializer = ActorMaterializer()
 
   private def selectSuitableTeam(user: User, dataSet: DataSet)(implicit ctx: DBAccessContext): Fox[ObjectId] =
     (for {
@@ -142,7 +142,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
       withFallback: Boolean,
       organizationName: String,
       oldTracingId: Option[String] = None)(implicit ctx: DBAccessContext): Fox[(Option[String], Option[String])] = {
-    def getFallbackLayer(): Option[SegmentationLayer] =
+    def getFallbackLayer: Option[SegmentationLayer] =
       if (withFallback) {
         dataSource.dataLayers.flatMap {
           case layer: SegmentationLayer => Some(layer)
@@ -169,7 +169,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
       case TracingType.volume =>
         for {
           client <- tracingStoreService.clientFor(dataSet)
-          fallbackLayer = getFallbackLayer()
+          fallbackLayer = getFallbackLayer
           _ <- bool2Fox(fallbackLayer.forall(_.elementClass != ElementClass.uint64)) ?~> "annotation.volume.uint64"
           volumeTracing <- createVolumeTracing(dataSource, organizationName, fallbackLayer)
           volumeTracingId <- client.saveVolumeTracing(volumeTracing)
@@ -177,7 +177,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
       case TracingType.hybrid =>
         for {
           client <- tracingStoreService.clientFor(dataSet)
-          fallbackLayer = getFallbackLayer()
+          fallbackLayer = getFallbackLayer
           _ <- bool2Fox(fallbackLayer.forall(_.elementClass != ElementClass.uint64)) ?~> "annotation.volume.uint64"
           skeletonTracingId <- client.saveSkeletonTracing(
             SkeletonTracingDefaults.createInstance.copy(dataSetName = dataSet.name,
@@ -221,7 +221,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
       implicit ctx: DBAccessContext): Fox[Unit] = {
     def createNewTracings(dataSet: DataSet, dataSource: DataSource) = annotation.tracingType match {
       case TracingType.skeleton =>
-        createTracingsForExplorational(dataSet, dataSource, TracingType.volume, true, organizationName).flatMap {
+        createTracingsForExplorational(dataSet, dataSource, TracingType.volume, withFallback = true, organizationName).flatMap {
           case (_, Some(volumeId)) => annotationDAO.updateVolumeTracingId(annotation._id, volumeId)
           case _                   => Fox.failure("unexpectedReturn")
         }
@@ -229,7 +229,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
         createTracingsForExplorational(dataSet,
                                        dataSource,
                                        TracingType.skeleton,
-                                       false,
+                                       withFallback = false,
                                        organizationName,
                                        annotation.volumeTracingId).flatMap {
           case (Some(skeletonId), _) => annotationDAO.updateSkeletonTracingId(annotation._id, skeletonId)
@@ -296,21 +296,15 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
     }).flatten
   }
 
-  def baseForTask(taskId: ObjectId)(implicit ctx: DBAccessContext): Fox[Annotation] =
+  private def baseForTask(taskId: ObjectId)(implicit ctx: DBAccessContext): Fox[Annotation] =
     (for {
       list <- annotationDAO.findAllByTaskIdAndType(taskId, AnnotationType.TracingBase)
     } yield list.headOption.toFox).flatten
 
-  def annotationsFor(taskId: ObjectId)(implicit ctx: DBAccessContext) =
+  def annotationsFor(taskId: ObjectId)(implicit ctx: DBAccessContext): Fox[List[Annotation]] =
     annotationDAO.findAllByTaskIdAndType(taskId, AnnotationType.Task)
 
-  def countOpenNonAdminTasks(user: User)(implicit ctx: DBAccessContext) =
-    for {
-      teamManagerTeamIds <- userService.teamManagerTeamIdsFor(user._id)
-      result <- annotationDAO.countActiveAnnotationsFor(user._id, AnnotationType.Task, teamManagerTeamIds)
-    } yield result
-
-  def tracingFromBase(annotationBase: Annotation, dataSet: DataSet)(
+  private def tracingFromBase(annotationBase: Annotation, dataSet: DataSet)(
       implicit ctx: DBAccessContext,
       m: MessagesProvider): Fox[(Option[String], Option[String])] =
     for {
@@ -415,7 +409,8 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
       )
     } yield volumeTracing
 
-  def abortInitializedAnnotationOnFailure(initializingAnnotationId: ObjectId, insertedAnnotationBox: Box[Annotation]) =
+  def abortInitializedAnnotationOnFailure(initializingAnnotationId: ObjectId,
+                                          insertedAnnotationBox: Box[Annotation]): Fox[Unit] =
     insertedAnnotationBox match {
       case Full(_) => Fox.successful(())
       case _       => annotationDAO.abortInitializingAnnotation(initializingAnnotationId)
@@ -428,7 +423,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
       volumeTracingIdBox: Box[Option[String]],
       dataSetId: ObjectId,
       description: Option[String]
-  )(implicit ctx: DBAccessContext) =
+  )(implicit ctx: DBAccessContext): Fox[Unit] =
     for {
       task <- taskFox
       skeletonIdOpt <- skeletonTracingIdBox.toFox
@@ -445,7 +440,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
                                   description.getOrElse(""),
                                   typ = AnnotationType.TracingBase)
       _ <- annotationDAO.insertOne(annotationBase)
-    } yield true
+    } yield ()
 
   def createFrom(user: User,
                  dataSet: DataSet,
@@ -469,13 +464,14 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
       _ <- annotationDAO.insertOne(annotation)
     } yield annotation
 
-  def sharedAnnotationsFor(userTeams: List[ObjectId])(implicit ctx: DBAccessContext) =
+  def sharedAnnotationsFor(userTeams: List[ObjectId])(implicit ctx: DBAccessContext): Fox[List[Annotation]] =
     sharedAnnotationsDAO.findAllSharedForTeams(userTeams)
 
-  def updateTeamsForSharedAnnotation(annotationId: ObjectId, teams: List[ObjectId])(implicit ctx: DBAccessContext) =
+  def updateTeamsForSharedAnnotation(annotationId: ObjectId, teams: List[ObjectId])(
+      implicit ctx: DBAccessContext): Fox[Unit] =
     sharedAnnotationsDAO.updateTeamsForSharedAnnotation(annotationId, teams)
 
-  def sharedTeamsFor(annotationId: ObjectId)(implicit ctx: DBAccessContext) =
+  def sharedTeamsFor(annotationId: ObjectId)(implicit ctx: DBAccessContext): Fox[List[Team]] =
     for {
       teamIds <- sharedAnnotationsDAO.sharedTeamsFor(annotationId)
       teamIdsValidated <- Fox.serialCombined(teamIds)(ObjectId.parse(_))
@@ -515,7 +511,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
   private def getTracingsScalesAndNamesFor(annotations: List[Annotation], skipVolumeData: Boolean)(
       implicit ctx: DBAccessContext): Fox[List[List[DownloadAnnotation]]] = {
 
-    def getSingleDownloadAnnotation(annotation: Annotation, dataSetId: ObjectId, scaleOpt: Option[Scale]) =
+    def getSingleDownloadAnnotation(annotation: Annotation, scaleOpt: Option[Scale]) =
       for {
         user <- userService.findOneById(annotation._user, useCache = true) ?~> "user.notFound"
         taskOpt <- Fox.runOptional(annotation._task)(taskDAO.findOne) ?~> "task.notFound"
@@ -565,8 +561,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
           skeletonTracings <- getSkeletonTracings(dataSetId, annotations.map(_.skeletonTracingId))
           volumeTracings <- getVolumeTracings(dataSetId, annotations.map(_.volumeTracingId))
           volumeDataObjects <- getVolumeDataObjects(dataSetId, annotations.map(_.volumeTracingId))
-          incompleteDownloadAnnotations <- Fox.serialCombined(annotations)(
-            getSingleDownloadAnnotation(_, dataSetId, scale))
+          incompleteDownloadAnnotations <- Fox.serialCombined(annotations)(getSingleDownloadAnnotation(_, scale))
         } yield
           incompleteDownloadAnnotations
             .zip(skeletonTracings)
@@ -596,7 +591,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
 
     def addToZip(nmls: List[(Enumerator[Array[Byte]], String, Option[Array[Byte]])]): Future[Boolean] =
       nmls match {
-        case (nml, name, volumeDataOpt) :: tail => {
+        case (nml, name, volumeDataOpt) :: tail =>
           if (volumeDataOpt.isDefined) {
             val subZip = temporaryFileCreator.create(normalize(name), ".zip")
             val subZipper =
@@ -611,7 +606,6 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
           } else {
             zipper.addFileFromEnumerator(name + ".nml", nml).flatMap(_ => addToZip(tail))
           }
-        }
         case _ =>
           Future.successful(true)
       }
@@ -632,25 +626,25 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
       updated <- annotationInformationProvider.provideAnnotation(typ, id, issuingUser)
     } yield updated
 
-  def resetToBase(annotation: Annotation)(implicit ctx: DBAccessContext, m: MessagesProvider) = annotation.typ match {
-    case AnnotationType.Explorational =>
-      Fox.failure("annotation.revert.tasksOnly")
-    case AnnotationType.Task =>
-      for {
-        task <- taskFor(annotation)
-        _ = logger.warn(
-          s"Resetting annotation ${annotation._id} to base, discarding skeleton tracing ${annotation.skeletonTracingId} and/or volume tracing ${annotation.volumeTracingId}")
-        annotationBase <- baseForTask(task._id)
-        dataSet <- dataSetDAO.findOne(annotationBase._dataSet)(GlobalAccessContext) ?~> "dataSet.notFoundForAnnotation"
-        (newSkeletonIdOpt, newVolumeIdOpt) <- tracingFromBase(annotationBase, dataSet)
-        _ <- Fox
-          .bool2Fox(newSkeletonIdOpt.isDefined || newVolumeIdOpt.isDefined) ?~> "annotation.needsEitherSkeletonOrVolume"
-        _ <- Fox.runOptional(newSkeletonIdOpt)(newSkeletonId =>
-          annotationDAO.updateSkeletonTracingId(annotation._id, newSkeletonId))
-        _ <- Fox.runOptional(newVolumeIdOpt)(newVolumeId =>
-          annotationDAO.updateVolumeTracingId(annotation._id, newVolumeId))
-      } yield ()
-  }
+  def resetToBase(annotation: Annotation)(implicit ctx: DBAccessContext, m: MessagesProvider): Fox[Unit] =
+    annotation.typ match {
+      case AnnotationType.Explorational =>
+        Fox.failure("annotation.revert.tasksOnly")
+      case AnnotationType.Task =>
+        for {
+          task <- taskFor(annotation)
+          _ = logger.warn(
+            s"Resetting annotation ${annotation._id} to base, discarding skeleton tracing ${annotation.skeletonTracingId} and/or volume tracing ${annotation.volumeTracingId}")
+          annotationBase <- baseForTask(task._id)
+          dataSet <- dataSetDAO.findOne(annotationBase._dataSet)(GlobalAccessContext) ?~> "dataSet.notFoundForAnnotation"
+          (newSkeletonIdOpt, newVolumeIdOpt) <- tracingFromBase(annotationBase, dataSet)
+          _ <- Fox.bool2Fox(newSkeletonIdOpt.isDefined || newVolumeIdOpt.isDefined) ?~> "annotation.needsEitherSkeletonOrVolume"
+          _ <- Fox.runOptional(newSkeletonIdOpt)(newSkeletonId =>
+            annotationDAO.updateSkeletonTracingId(annotation._id, newSkeletonId))
+          _ <- Fox.runOptional(newVolumeIdOpt)(newVolumeId =>
+            annotationDAO.updateVolumeTracingId(annotation._id, newVolumeId))
+        } yield ()
+    }
 
   private def settingsFor(annotation: Annotation)(implicit ctx: DBAccessContext) =
     if (annotation.typ == AnnotationType.Task || annotation.typ == AnnotationType.TracingBase)
@@ -669,7 +663,7 @@ class AnnotationService @Inject()(annotationInformationProvider: AnnotationInfor
   def publicWrites(annotation: Annotation,
                    requestingUser: Option[User] = None,
                    restrictionsOpt: Option[AnnotationRestrictions] = None): Fox[JsObject] = {
-    implicit val ctx = GlobalAccessContext
+    implicit val ctx: DBAccessContext = GlobalAccessContext
     for {
       dataSet <- dataSetDAO.findOne(annotation._dataSet) ?~> "dataSet.notFoundForAnnotation"
       organization <- organizationDAO.findOne(dataSet._organization) ?~> "organization.notFound"
